// src/api/types.d.ts
export interface ServiceCycle {
  id:            number;
  cycleNumber:   number;
  pricePerCycle: number;
}

export interface ServiceStep {
  id:                      number;
  stepNo:                  number;
  name:                    string;
  estimatedCompletionTime: string;
  description:             string;
  documentRequired:        DocumentRequired[];
  result:                  Result[];
}

export interface DocumentRequired {
  id:               number;
  documentRequired: string;
}

export interface Result {
  id:     number;
  result: string;
}
// ...
export type ApiMasterServiceParam = {
  appliedNation:      string[];
  appliedCompanyType: string[];
  serviceType:        string;
  serviceName:        string;
  serviceDescription: string;
  serviceStep:        ServiceStep[];
  serviceCycle:       ServiceCycle[];
}